[CACHE] Rename zend cache handler to Cache\Core\ZendCacheHandler
